added detect model name based on model code
